:recycle: refactor(#168/user): user 관련 서비, dto 클래스 ninedot-application 모듈로 이동
