+++ ninedot-application/src/main/java/org/sopt36/ninedotserver/user/usecase/UserQueryService.java
@@ -1,12 +1,12 @@
-package org.sopt36.ninedotserver.user.service.query;
+package org.sopt36.ninedotserver.user.usecase;
 
 import static org.sopt36.ninedotserver.user.exception.UserErrorCode.USER_NOT_FOUND;
 
 import lombok.RequiredArgsConstructor;
-import org.sopt36.ninedotserver.user.domain.User;
 import org.sopt36.ninedotserver.user.dto.response.UserInfoResponse;
 import org.sopt36.ninedotserver.user.exception.UserException;
-import org.sopt36.ninedotserver.user.repository.UserRepository;
+import org.sopt36.ninedotserver.user.model.User;
+import org.sopt36.ninedotserver.user.port.out.UserRepositoryPort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,7 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class UserQueryService {
 
-    private final UserRepository userRepository;
+    private final UserRepositoryPort userRepository;
 
     public UserInfoResponse getUserInfo(Long userId) {
         User user = userRepository.findById(userId)
